Result commit

+++ PhotoEditor/IImageSaver.cs
@@ -2,5 +2,5 @@
 
 public interface IImageSaver
 {
-    void Save(IImage image, int dpiX, int dpiY, string path);
+    void Save(IImage image, float dpiX, float dpiY, string path);
 }
