+++ PhotoEditor/IPixelMapUploader.cs
@@ -4,5 +4,5 @@ namespace PhotoEditor;
 
 public interface IPixelMapUploader
 {
-    BitmapSource Upload(byte[] date, int width, int height, int dpiX, int dpiY);
+    BitmapSource Upload(byte[] date, int width, int height, float dpiX, float dpiY);
 }
